backend/requirements.txt
# Original reqs
# flask
# flask-cors
# python-dotenv
# itsdangerous
# werkzeug
# flask-jwt-extended
# pytz
# stripe
# sqlalchemyflask_sqlalchemy

# NEW reqs
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
itsdangerous==2.1.2
werkzeug==3.0.1
flask-jwt-extended==4.6.0
pytz==2024.1
stripe==7.10.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
qrcode[pil]==7.4.2
pillow==10.2.0